Exclude poorly performing FlowGraphAction from action proxy results
